Move BrokerState to package mqtt.broker .
